css edits. ready to merge with master and turn in

+++ app.js
@@ -98,7 +98,7 @@ function handleSubmit(event){
   var storeName = event.target.cookieStoreName.value;// event=eventlistener, target=target node"storeFormEl", cookieStoreNeme node within target node, value is input of node
   var minCustomers = parseInt(event.target.minCust.value); //gets minCust value, parseInt prevents input of string
   var maxCustomers = parseInt(event.target.maxCust.value);
-  var avgCookies = parseInt(event.target.avgCookies.value);
+  var avgCookies = parseFloat(event.target.avgCookies.value);
 
   // console.log(storeName);
   // console.log(minCustomers);
